requirements.txt
flask>=3.1.2
pandas>=3.0.0
openpyxl>=3.1.5
requests>=2.32.5
pytest>=9.0.2
pytest-cov>=7.0.0
pydantic-ai[openrouter]>=1.48.0
schedule>=1.2.2
flasgger>=0.9.7.1
google-auth>=2.48.0
google-api-python-client>=2.188.0
yoyo-migrations>=9.0.0
pypdf>=6.6.2
